requirements/base.txt
psycopg2==2.6.1
Django==1.9.5
djangorestframework==3.3.3
django-rest-swagger==0.3.5
drf-extensions==0.2.8
django-redis==4.3.0
redis==2.10.5
